+++ src/fraction_operations.c
@@ -6,7 +6,7 @@
 /*   By: rjw <rjw@student.codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/03/12 02:52:56 by rjw           #+#    #+#                 */
-/*   Updated: 2025/03/29 03:51:38 by rjw           ########   odam.nl         */
+/*   Updated: 2025/04/11 14:17:47 by rjw           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*populate_numerator(char *num_str, uint64_t mant, int16_t expo)
 	char	exp_str[MAX_DBL_STR_LEN + 1];
 	size_t	mant_len;
 
-	mant_len = int64_base(mant, DECIMAL_BASE, mant_bits, sizeof(mant_bits));
+	mant_len = int64_base((int64_t)mant, DECIMAL_BASE, mant_bits, sizeof(mant_bits));
 	if (mant_len >= MAX_DBL_STR_LEN)
 		return (NULL);
 	if (expo > 0)
